feat(validator): add nil machine claim error
All epochs have at least one input. If an input has no outputs, then Input.OutputsHash equals the last input OutputsHash. If it is the first input and it has no outputs, its OutputsHash is the application's template hash. Therefore, if this value is nil the node is in an inconsistent state.