fix(ethutil): handle rpc.HTTPError on queryBlockRangeTooLarge
This is a workaround for providers that break JSON-RPC on error
handling. Some of them return HTTP errors when they are unable to
fulfill the request such as a query with a large block range as if this
was a REST API.

So now when we test for queryBlockRangeTooLarge, we also unwrap HTTP
errors assuming it may still JSON contents, and use that for our query.