feat(evmreader): add a fast inputs sync
Previously the node had to scan the blockchain logs from input box
deployment up to the latest block to find all inputs of a new
application. This new feature implements an optimization that skips this
scanning when the application has no inputs and does a binary search to
find the exact blocks where the inputs happened. The objective is to
reduce provider usage and reduce the time it takes to bring new
applications online.

On the implementation side, we've added a binary search utility function
that caches the partial search ranges from one input to the next to
further reduce provider usage instead of multiple individual searches.

The function accepts any f, x, count such that f(x) -> count.
MBSearch finds the xs where count has changed. Concretely, x is a block
number, count is the number of inputs and f is GetNumberOfInputs.